requirements.txt
llama-cpp-python==0.2.85
huggingface-hub==0.24.0
Pillow==10.3.0
piexif==1.1.3
tqdm==4.66.0
requests==2.32.0
